requirement.txt
# FastAPI framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database & ORM
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9   # Only needed if using PostgreSQL
# If only using SQLite, you can remove psycopg2-binary

# Data validation & settings
pydantic>=2.8.0
pydantic-settings>=2.3.0

# Authentication & security
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.9
bcrypt>=4.0.1

# Email support
fastapi-mail>=1.4.1
email-validator>=2.1.0

# External services
httpx>=0.27.0
pybreaker>=0.8.0

# QR code generation
qrcode[pil]>=7.4.2

# Environment variables
python-dotenv>=1.0.1
